bootcamps_cursos/Biblioteca/requirements.txt

psycopg2-binary==2.9.9
openai==1.14.3
matplotlib==3.8.4
python-dotenv==1.0.1
pandas==2.3.0
numpy==2.3.0